Delete test-cases version_declaration-003 and -004.
In https://www.w3.org/Bugs/Public/show_bug.cgi?id=14955#c10,
I said that the proper response to comment 5 (which pointed out that
$input-context is undefined in these queries) would have been to repair
the queries, not change their expectations.

I still think that's true, but I now see that if we were to repair the
queries, they would become near-duplicates of prolog-version-6 amd -7,
with the only significant difference in the <test> being the presence
(in vd-3,4) of the comment (:encoding "abc":) before the version declaration.
But other test-cases in this test-set cover that possibility more
straightforwardly, so vd-3 & -4 wouldn't really add anything.